Декомпозиция кода
