+++ lab_4/constants.py
@@ -1,9 +1,9 @@
 class Constants:
-    # GLOBALS VARS
+
     SCREEN_WIDTH = 800
     SCREEN_HEIGHT = 700
-    PLAY_WIDTH = 300  # meaning 300 // 10 = 30 width per block
-    PLAY_HEIGHT = 600  # meaning 600 // 20 = 20 height per block
+    PLAY_WIDTH = 300  # 300 // 10 = 30 width per block
+    PLAY_HEIGHT = 600  # 600 // 20 = 20 height per block
     BLOCK_SIZE = 30
 
     TOP_LEFT_X = (SCREEN_WIDTH - PLAY_WIDTH) // 2
@@ -142,6 +142,3 @@ class Constants:
     SCORE_PER_ROW = 10
 
     GAME_OVER_DELAY_MS = 2000
-
-
-
